refactor: foreign key and active flag in ClientVerfuegung
In a test with about 1'500'000 records in table clientverfuegung the
performance dropped significantly when joining to the client table
and filtering with the active flag in the client table.

To improve the performance, it was decided to synchronise the active
flag from client to clientverfuegung with the client_active_toggle
trigger.
With this redundancy, joining the client table is not necessary.