chore: use TestContainers KafkaContainer
When using Kafka dirrectly with docker-compose in the test environment,
there is a configuration error, because the Kafka container must start
with the property KAFKA_ADVERTISED_LISTENERS, to advertise itself.
Before starting the test environment, the correct value is not known,
because Testcontainers assigns the port randomly, thus the consumers
will always assume that the group coordinator is reachable on
localhost:9092 (the default values).

Maybe it is possible to change the property while the Kafka container
is running, but there is an easier solution: Testcontainers provides
a configuration for a Kafka container, which makes it easy to get the
correct bootstrap server URL.