refactor: set Kafka consumer group id to pom version
We want to make sure we can deploy multiple versions of the exchange
service, reading all the data from Kafka, thus we must make sure they
are running in different groups.

With this approach we cannot run multiple exchange services with the
same version!
To support that use case, we have to wait until Quarkus allows the
mp.* configuration to be set at runtime, or use an undefined group.id
as we do for development mode.

The disadvantage of the latter is, that each restart of the service
will will generate a new group.id, trigger Kafka to re-balance and will
make the service to read ALL events again.