fix: exchange-service requires group.id property
Thanks to the new quarkus version, the group.id property finally works
as intended: a restart of the exchange-service with the same group.id
property will join the Kafka group at the latest committed offset and thus
only read new events.

An exchange-service with a new group.id property will read ALL events
from Kafka. This is a good case to launch an exchange-service with a new
(incompatible) API.