chore: windows compatible setup
The set-secrets.sh script didn't worked on Windows,
because it is typically checked-out with CRLF line endings.
As a result a file "kibon-realm.json\r" was created in the container
and thus the file "kibon-realm.json" not found.

Also updated the readme to specify the docker-compose project directory,
which is necessary to get the environment variables form the docker/.env
file applied when calling docker-compose from the project root.