KIBON-1653 refactor: filter Verfuegungen without any Zeitabschnitte
This is kind of redundant, since a Verfuegung contains Zeitabschnitte for
each month of a period. Since Verfuegungen outside the accessible range
are already filtered on database level, there should in practice never
be a Verfuegung without Zeitabschnitte.

The removeIf condition for Verfuegung is quite simple to implement though,
and the additional unit tests help catching cases that are harderd to
setup up as integration tests.