refactor: introduce a BaseEntity
Starting with Quarkus 2.0, there is a class loading issue
with our json TypeDefs. An entity class is with json-node
type is loaded before the TypeDefs class is loaded, resulting
in a jsonb-node class not found exception.

The issue is resolved by ensuring that the type def is loaded
first, by using a BaseEntity from which the other entities are
actually extending.