an attempt to fix the "no eventId header" problem
Since https://github.com/smallrye/smallrye-reactive-messaging/pull/1272, there
are two different OutgoingKafkaRecordMetadata classes.

When calling KafkaRecord.of(key, payload).addMetadata(someMetadata) a message is
generated which actually contains 2 metadata objects. One for each metadata class.

My suspicion is, that for some reason once in a while the wrong metadata object
is passed on to Kafka. Maybe Quarkus' loader is not entirely deterministic?
It's a wild guess...