disable quarkus devservices in tests - we handle versions manually
On CI (after a clean checkout of the repository), the environment is
not properly setup. There is no .env file, only a .env-template. This
makes quarkus miss configuration properties for keycloak, postgres and kafka
such that different versions are started than the ones used by the
project.

Since we manage versions manually, the devservices are disabled for the
test profile.